* Change statesInMem to use Redis protocol * SocketIo is still there as renamed class * remove deprecated "new Buffer()" usage (just came over it while testing) * Add redis and respjs to dependecies ...
Open:
* think about a "client side fallback logic" if server is still on socketio to add a auto fallback
* also convert bjects
