+++ lib/adapter.js
@@ -283,7 +283,7 @@ function Adapter(options) {
     let States;
     if (config.states && config.states.type) {
         if (config.states.type === 'file') {
-            States = require('./states/statesInMemClient');
+            States = require('./states/statesInRedis');
         } else if (config.states.type === 'redis') {
             States = require('./states/statesInRedis');
         } else {
@@ -5333,7 +5333,7 @@ function Adapter(options) {
                         for (let i = 0; i < keys.length; i++) {
                             // We can JSON.parse, but index is 16x faster
                             if (obj[i]) {
-                                const id = keys[i].substring(0, keys[i].length - '.logging'.length).replace(/^io\./, '');
+                                const id = keys[i].substring(0, keys[i].length - '.logging'.length);
                                 if ((typeof obj[i] === 'string' && (obj[i].indexOf('"val":true') !== -1 || obj[i].indexOf('"val":"true"') !== -1)) ||
                                     (typeof obj[i] === 'object' && (obj[i].val === true || obj[i].val === 'true'))) {
                                     logs.push('Subscriber - ' + id + ' ENABLED');
