+++ lib/inMemFileDB.js
@@ -298,10 +298,12 @@ class InMemoryFileDB {
         }
 
         const clients = this.getClients();
+        let publishCount = 0;
 
         for (const i in clients) {
             if (clients.hasOwnProperty(i)) {
-                this.publishToClients(clients[i], type, id, obj);
+                // @ts-ignore
+                publishCount += this.publishToClients(clients[i], type, id, obj);
             }
         }
 
@@ -315,6 +317,7 @@ class InMemoryFileDB {
                 }
             }
         }
+        return publishCount;
     }
 
     // Destructor of the class. Called by shutting down.
