+++ lib/objects/objectsUtils.js
@@ -121,7 +121,7 @@ function WMStrm(key, options) {
 
     Writable.call(this, options); // init super
     this.key = key; // save key
-    memStore[key] = new Buffer(''); // empty
+    memStore[key] = Buffer.alloc(0); // empty
 }
 util.inherits(WMStrm, Writable);
 
@@ -130,11 +130,11 @@ WMStrm.prototype._write = function (chunk, enc, cb) {
         // our memory store stores things in buffers
         const buffer = (Buffer.isBuffer(chunk)) ?
             chunk :  // already is Buffer use it
-            new Buffer(chunk, enc);  // string, convert
+            Buffer.from(chunk, enc);  // string, convert
 
         // concatenate to the buffer already there
         if (!memStore[this.key]) {
-            memStore[this.key] = new Buffer('');
+            memStore[this.key] = Buffer.alloc(0);
             console.log('memstore for ' + this.key + ' is null');
         }
         memStore[this.key] = Buffer.concat([memStore[this.key], buffer]);
