+++ lib/setup.js
@@ -2071,6 +2071,7 @@ function getRepository(repoUrl, params, callback) {
                                 tools.getRepositoryFile(repos.native.repositories[active].link, {
                                     hash: repos.native.repositories[active].hash,
                                     sources: repos.native.repositories[active].json,
+                                    // @ts-ignore
                                     controller: require('../io-package').common.version,
                                     node: process.version,
                                     name: tools.appName
@@ -2210,7 +2211,7 @@ function dbConnect(onlyCheck, params, callback) {
         if (!isStatesConnected) {
             if (config.states.type === 'file') {
                 // Just open in memory DB itself
-                States = require('./states/statesInMemServer');
+                States = require('./states/statesInMemServerRedis');
                 states = new States({
                     connection: config.states,
                     logger: {
