+++ lib/states.js
@@ -5,7 +5,7 @@ const config = JSON.parse(require('fs').readFileSync(getConfigFileName(), 'utf8'
 if (!config.states) config.states = {type: 'file'};
 
 if (config.states.type === 'file') {
-    module.exports = require('./states/statesInMemClient');
+    module.exports = require('./states/statesInRedis');
 } else if (config.states.type === 'redis') {
     module.exports = require('./states/statesInRedis');
 } else {
