+++ lib/states/statesInRedis.js
@@ -484,7 +484,7 @@ function StateRedis(settings) {
         if (!clientBin) _createBinaryClient ();
         clientBin.get(id, (err, data) => {
             if (!err && data) {
-                if (callback) callback(err, new Buffer(data, 'binary'));
+                if (callback) callback(err, Buffer.from(data, 'binary'));
             } else {
                 if (callback) callback(err);
             }
